Shrink: int wallSmooth counter and MAX_WALL_SMOOTH constant
wallSmooth returns int, uses int loop counter. MAX_WALL_SMOOTH
changed to static final int. Int comparisons and increments are
cheaper bytecode than double. -5 bytes (1472 → 1467).

+++ src/pez/mini/Pugilist.java
@@ -24,7 +24,7 @@ public class Pugilist extends AdvancedRobot {
     static final double WALL_MARGIN = 20;
     static final double BULLET_POWER = 1.9;
     static final double MAX_BULLET_POWER = 3.0;
-    static final double MAX_WALL_SMOOTH = 97;
+    static final int MAX_WALL_SMOOTH = 97;
     static final double BOT_WIDTH = 40;
 
     static Rectangle2D fieldRectangle = new Rectangle2D.Double(WALL_MARGIN, WALL_MARGIN,
@@ -141,9 +141,9 @@ public class Pugilist extends AdvancedRobot {
     }
 
     static Point2D wallSmoothedDestination(Point2D location, double direction) {
-        double s = wallSmooth(location, enemyLocation, direction);
+        int s = wallSmooth(location, enemyLocation, direction);
         if (s >= MAX_WALL_SMOOTH - 1) {
-            double rs = wallSmooth(location, enemyLocation, -direction);
+            int rs = wallSmooth(location, enemyLocation, -direction);
             if (rs < s) {
                 direction = -direction;
                 s = rs;
@@ -157,8 +157,8 @@ public class Pugilist extends AdvancedRobot {
                 - direction * (Math.PI / 2 + 0.2 - (w / 100.0)), enemyDistance / 5);
     }
 
-    static double wallSmooth(Point2D from, Point2D toward, double direction) {
-        double w = 0;
+    static int wallSmooth(Point2D from, Point2D toward, double direction) {
+        int w = 0;
         while (w < MAX_WALL_SMOOTH && !fieldRectangle.contains(orbitProject(from, toward, direction, w++)))
             ;
         return w;
